Add files via upload
In this project, I made a menu where, when clicked, it goes to details and shows a bigger picture of the menu item. I reduced the opacity of the images so that the text is easier to see. Also, the details page has an elevated button to return to the homepage. While navigating through the pages, it doesn't stack. I used navigation replacement so the pages doesn't stack up.